+++ sdk/core/azure-core/azure/core/pipeline/policies/_utils.py
@@ -26,29 +26,7 @@
 import datetime
 import email.utils
 from requests.structures import CaseInsensitiveDict
-
-class _FixedOffset(datetime.tzinfo):
-    """Fixed offset in minutes east from UTC.
-
-    Copy/pasted from Python doc
-
-    :param int offset: offset in minutes
-    """
-
-    def __init__(self, offset):
-        self.__offset = datetime.timedelta(minutes=offset)
-
-    def utcoffset(self, dt):
-        return self.__offset
-
-    def tzname(self, dt):
-        return str(self.__offset.total_seconds()/3600)
-
-    def __repr__(self):
-        return "<FixedOffset {}>".format(self.tzname(None))
-
-    def dst(self, dt):
-        return datetime.timedelta(0)
+from ..._utils import _FixedOffset
 
 def _parse_http_date(text):
     """Parse a HTTP date format into datetime."""
